refactor: 결과 패키지를 통계 패키지로 변경
